Move to all tests in the cwd with separate Terraform dirs
