+++ test/test_bastion.py
@@ -6,18 +6,13 @@ import botocore.exceptions
 import pytest
 from attrs import define
 
-from ..utils import MotoServer, Terraform, TerraformOutputs
+from .utils import MotoServer, Terraform, TerraformOutputs
 
 
 @define
 class Outputs(TerraformOutputs):
     id: str
     public_ip: str
-
-
-@pytest.fixture(scope="module")
-def this_dir() -> Path:
-    return Path(__file__).parent
 
 
 @pytest.fixture(scope="module", autouse=True)
@@ -39,8 +34,12 @@ def key_pair(ec2) -> ...:
 
 @pytest.fixture(scope="module")
 def base_vars(subnet, key_pair) -> dict[str, Any]:
+    # This AMI should exist in the Moto server.
+    # Refer to https://github.com/getmoto/moto/blob/master/moto/ec2/resources/amis.json.
+    ami = "ami-12c6146b"
+
     return {
-        "ami": "ami-12c6146b",
+        "ami": ami,
         "key_name": key_pair.key_name,
         "subnet_id": subnet.id,
     }
@@ -48,20 +47,11 @@ def base_vars(subnet, key_pair) -> dict[str, Any]:
 
 @pytest.fixture(scope="module")
 def tf(this_dir: Path, moto_server) -> Terraform:
-    tf = Terraform(this_dir, f"http://localhost:{moto_server.port}")
+    tf = Terraform(
+        this_dir / "bastion", f"http://localhost:{moto_server.port}"
+    )
     tf.init(upgrade=True)
     return tf
-
-
-@pytest.fixture(scope="module", autouse=True)
-def apply(tf: Terraform, base_vars: dict[str, Any]) -> None:
-    tf.apply(vars=base_vars)
-
-
-@pytest.fixture(scope="module")
-def instance(ec2, tf: Terraform):
-    outputs = Outputs.from_terraform(tf)
-    return ec2.Instance(outputs.id)
 
 
 @pytest.fixture(scope="module", autouse=True)
@@ -71,33 +61,19 @@ def reset(moto_server: MotoServer, this_dir: Path) -> None:
     (this_dir / "terraform.tfstate").unlink(missing_ok=True)
 
 
-def test_instance_exists(instance):
+def test_defaults(base_vars: dict[str, Any], tf: Terraform):
+    tf.apply(vars=base_vars)
+    outputs = Outputs.from_terraform(tf)
+    instance = ec2.Instance(outputs.id)
+
     try:
         instance.load()
     except botocore.exceptions.ClientError as exc:
         raise AssertionError from exc
 
-
-def test_ami(base_vars, instance):
     assert instance.image_id == base_vars["ami"]
-
-
-def test_instance_type(base_vars, instance):
     assert instance.instance_type == "t3.nano"
-
-
-def test_key_name(base_vars, instance):
     assert instance.key_name == base_vars["key_name"]
-
-
-def test_public_ip_address(tf, instance):
-    outputs = Outputs.from_terraform(tf)
     assert instance.public_ip_address == outputs.public_ip
-
-
-def test_security_groups(instance):
     assert len(instance.security_groups) == 1
-
-
-def test_subnet_id(base_vars, instance):
     assert instance.subnet_id == base_vars["subnet_id"]
