+++ test/test_eks.py
@@ -2,21 +2,40 @@ import uuid
 from pathlib import Path
 from typing import Any
 
-import boto3
+import botocore.errorfactory
 import pytest
+from attrs import define
 
-from ..utils import MotoServer, Terraform
-from . import Cluster
-
-
-@pytest.fixture(scope="module")
-def eks_client(moto_server: MotoServer) -> ...:
-    return boto3.client("eks", endpoint_url=moto_server.endpoint)
+from .utils import MotoServer, Terraform
 
 
-@pytest.fixture(scope="module")
-def this_dir() -> Path:
-    return Path(__file__).parent
+@define
+class Cluster:
+    name: str
+    version: str
+    endpoint_public_access: bool
+    endpoint_private_access: bool
+    security_group_ids: set[str]
+
+    @classmethod
+    def from_name(cls, eks_client: ..., name: str):
+        try:
+            resp = eks_client.describe_cluster(name=name)
+        except botocore.errorfactory.ResourceNotFoundException:
+            return None
+
+        if resp["ResponseMetadata"]["HTTPStatusCode"] != 200:
+            return None
+
+        cluster = resp["cluster"]
+
+        return cls(
+            name,
+            cluster["version"],
+            cluster["resourcesVpcConfig"]["endpointPublicAccess"],
+            cluster["resourcesVpcConfig"]["endpointPrivateAccess"],
+            set(cluster["resourcesVpcConfig"]["securityGroupIds"]),
+        )
 
 
 @pytest.fixture(scope="module", autouse=True)
@@ -60,7 +79,7 @@ def base_vars(subnets):
 
 @pytest.fixture(scope="module")
 def tf(this_dir: Path, moto_server) -> Terraform:
-    tf = Terraform(this_dir, f"http://localhost:{moto_server.port}")
+    tf = Terraform(this_dir / "eks", f"http://localhost:{moto_server.port}")
     tf.init(upgrade=True)
     return tf
 
@@ -75,22 +94,13 @@ def test_defaults(eks_client: ..., base_vars: dict[str, Any], tf: Terraform):
     tf.apply(vars=base_vars)
     cluster = Cluster.from_name(eks_client, base_vars["name"])
     assert cluster
-    assert cluster.version == base_vars["cluster_version"]
     assert cluster.endpoint_private_access
     assert cluster.endpoint_public_access
     assert len(cluster.security_group_ids) == 0
+    assert cluster.version == base_vars["cluster_version"]
 
 
-def test_name(eks_client: ..., tf: Terraform, base_vars: dict[str, Any]):
-    name = str(uuid.uuid4())
-    tf.apply(vars=base_vars | {"name": name})
-    cluster = Cluster.from_name(eks_client, name)
-    assert cluster.name == name
-
-
-@pytest.mark.parametrize(
-    "cluster_version", ("1.23", "1.24", "1.25", "1.26", "1.27")
-)
+@pytest.mark.parametrize("cluster_version", ("1.23", "1.24", "1.25", "1.26"))
 def test_cluster_version(
     eks_client: ...,
     tf: Terraform,
@@ -102,20 +112,20 @@ def test_cluster_version(
     assert cluster.version == cluster_version
 
 
-def test_endpoint_public_access(
-    eks_client: ..., tf: Terraform, base_vars: dict[str, Any]
-):
-    tf.apply(vars=base_vars | {"endpoint_public_access": False})
-    cluster = Cluster.from_name(eks_client, base_vars["name"])
-    assert not cluster.endpoint_public_access
-
-
 def test_endpoint_private_access(
     eks_client: ..., tf: Terraform, base_vars: dict[str, Any]
 ):
     tf.apply(vars=base_vars | {"endpoint_private_access": False})
     cluster = Cluster.from_name(eks_client, base_vars["name"])
     assert not cluster.endpoint_private_access
+
+
+def test_endpoint_public_access(
+    eks_client: ..., tf: Terraform, base_vars: dict[str, Any]
+):
+    tf.apply(vars=base_vars | {"endpoint_public_access": False})
+    cluster = Cluster.from_name(eks_client, base_vars["name"])
+    assert not cluster.endpoint_public_access
 
 
 def test_security_groups(
