+++ test/test_key_pair.py
@@ -6,15 +6,10 @@ from typing import Any
 import botocore.exceptions
 import pytest
 
-from ..utils import MotoServer, Terraform
+from .utils import MotoServer, Terraform
 
 
-@pytest.fixture(scope="package")
-def this_dir() -> Path:
-    return Path(__file__).parent
-
-
-@pytest.fixture(scope="package")
+@pytest.fixture(scope="module")
 def base_vars(this_dir: Path) -> dict[str, Any]:
     key_name = str(uuid.uuid4())
     return {
@@ -23,31 +18,28 @@ def base_vars(this_dir: Path) -> dict[str, Any]:
     }
 
 
-@pytest.fixture(scope="package")
+@pytest.fixture(scope="module")
 def tf(this_dir: Path, moto_server) -> Terraform:
-    tf = Terraform(this_dir, f"http://localhost:{moto_server.port}")
+    tf = Terraform(
+        this_dir / "key_pair", f"http://localhost:{moto_server.port}"
+    )
     tf.init(upgrade=True)
     return tf
 
 
-@pytest.fixture(scope="module", autouse=True)
-def apply(tf: Terraform, base_vars: dict[str, Any]) -> None:
-    tf.apply(vars=base_vars)
-
-
-@pytest.fixture(scope="module", autouse=True)
+@pytest.fixture(autouse=True)
 def reset(moto_server: MotoServer, this_dir: Path) -> None:
     yield
     moto_server.reset()
     (this_dir / "terraform.tfstate").unlink(missing_ok=True)
 
 
-def test_key_pair_exists(ec2, base_vars: dict[str, Any]):
+def test_defaults(ec2: ..., base_vars: dict[str, Any], tf: Terraform):
+    tf.apply(vars=base_vars)
+
     try:
         ec2.KeyPair(base_vars["key_name"]).load()
     except botocore.exceptions.ClientError as exc:
         raise AssertionError from exc
 
-
-def test_key_pair_written(base_vars: dict[str, Any]):
     assert os.path.exists(base_vars["filename"])
